src/chwimi/requirements.txt
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.1
django-allauth==0.39.1
gunicorn==19.9.0
idna==2.8
oauthlib==3.0.1
Pillow==6.0.0
psycopg2-binary==2.8.3
python3-openid==3.1.0
pytz==2019.1
recombee-api-client==2.4.0
requests==2.22.0
requests-oauthlib==1.2.0
selenium==3.141.0
sqlparse==0.3.0
urllib3==1.25.2
whitenoise==4.1.2
